Worker/requirements.txt
Flask==2.2.5
Pillow==10.2.0
numpy==1.26.4